Circuit Breaker: Row Filter Limits
Adds the concept of circuit breaking to any filtering command that
requires Cassandra to retain too many Rows on heap. For backwards
compatibility this defaults to Integer.MAX_VALUE so that no queries
are killed automatically.

If you want to enable this set circuit_breaker_row_filter_limit in the
cassandra.yaml or use jmx endpoints on StorageService.